// apps/pc/src/generated/models/Pl_partnersharesettingsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pl_partnersharesettingspl_accesslevelcode = {
  100000000: '閲覧',
  100000001: '編集'
} as const;
export type Pl_partnersharesettingspl_accesslevelcode = keyof typeof Pl_partnersharesettingspl_accesslevelcode;
export const Pl_partnersharesettingspl_principalkindcode = {
  100000000: '個人',
  100000001: 'Team'
} as const;
export type Pl_partnersharesettingspl_principalkindcode = keyof typeof Pl_partnersharesettingspl_principalkindcode;
export const Pl_partnersharesettingspl_settingstatuscode = {
  100000000: '有効',
  100000001: '取消'
} as const;
export type Pl_partnersharesettingspl_settingstatuscode = keyof typeof Pl_partnersharesettingspl_settingstatuscode;
export const Pl_partnersharesettingsstatecode = {
  0: 'アクティブ',
  1: '非アクティブ'
} as const;
export type Pl_partnersharesettingsstatecode = keyof typeof Pl_partnersharesettingsstatecode;
export const Pl_partnersharesettingsstatuscode = {
  1: 'アクティブ',
  2: '非アクティブ'
} as const;
export type Pl_partnersharesettingsstatuscode = keyof typeof Pl_partnersharesettingsstatuscode;

export interface Pl_partnersharesettingsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  pl_accesslevelcode: Pl_partnersharesettingspl_accesslevelcode;
  pl_contenthash: string;
  pl_ismanagedprojection: boolean;
  pl_isprotectedmanagementpath: boolean;
  pl_name?: string;
  "pl_PartnerLookup@odata.bind": string;
  pl_partnersharesettingid: string;
  pl_principalkindcode: Pl_partnersharesettingspl_principalkindcode;
  pl_requestkey: string;
  pl_settingstatuscode: Pl_partnersharesettingspl_settingstatuscode;
  "pl_TeamPrincipalLookup@odata.bind"?: string;
  "pl_UserPrincipalLookup@odata.bind"?: string;
  statecode: Pl_partnersharesettingsstatecode;
  statuscode?: Pl_partnersharesettingsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pl_partnersharesettings extends Pl_partnersharesettingsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pl_accesslevelcodename?: string;
  pl_ismanagedprojectionname?: string;
  pl_isprotectedmanagementpathname?: string;
  pl_partnerlookupname?: string;
  pl_principalkindcodename?: string;
  pl_settingstatuscodename?: string;
  pl_teamprincipallookupname?: string;
  pl_teamprincipallookupyominame?: string;
  pl_userprincipallookupname?: string;
  pl_userprincipallookupyominame?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pl_partnerlookup?: object;
  _pl_partnerlookup_value?: string;
  pl_teamprincipallookup?: object;
  _pl_teamprincipallookup_value?: string;
  pl_userprincipallookup?: object;
  _pl_userprincipallookup_value?: string;
}